POC2-keras/requirements.txt
tensorflow>=2.12.0
numpy>=1.19.2
Pillow>=8.0.0
pandas>=1.2.0
matplotlib>=3.7.0
keras-ocr>=0.9.3
scikit-learn>=1.2.0
tqdm>=4.65.0
opencv-python>=4.8.0